editor: Move breadcrumb navigation code into element/breadcrumbs.rs
Splits all breadcrumb path/symbol dropdown code (segment building, the
directory browser popover, the outline-item and symbol-segment renderers,
the sibling/child/top-level outline index helpers they share with
document_symbols.rs, and their tests) out of the already very large
element.rs into its own element/breadcrumbs.rs submodule, per maintainer
review on #60282 asking for this to be split out.

element.rs re-exports the handful of items still needed by editor.rs,
document_symbols.rs, and editor_tests.rs (render_breadcrumb_text stays pub
since editor.rs re-exports it as part of the crate's public API; the rest
are pub(crate)) so none of those call sites needed to change. Nothing in
the moved code was public outside the crate except render_breadcrumb_text,
which was already pub before the move.
